[REFACTOR] calculator, lotto 패키지에서 함께 사용하는 클래스를 common 패키지로 이동
