+++ src/main/java/common/ExpressionSplitter.java
@@ -1,4 +1,4 @@
-package calculator;
+package common;
 
 import java.util.Optional;
 import java.util.regex.Matcher;
