+++ src/main/java/common/NumberElementCollection.java
@@ -1,4 +1,4 @@
-package model;
+package common;
 
 import lotto.exception.OutOfCountException;
