+++ src/test/java/calculator/NumberElementTest.java
@@ -1,6 +1,6 @@
 package calculator;
 
-import model.NumberElement;
+import common.NumberElement;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
