+++ src/test/java/lotto/LottoTicketTest.java
@@ -3,15 +3,13 @@ package lotto;
 import lotto.exception.DuplicateNumberException;
 import lotto.exception.OutOfCountException;
 import lotto.exception.OutOfMaxNumberException;
-import model.NumberElement;
-import model.NumberElementCollection;
+import common.NumberElementCollection;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import java.util.List;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.*;
