+++ src/test/java/lotto/LottoWalletTest.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import model.NumberElementCollection;
+import common.NumberElementCollection;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -8,7 +8,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class LottoWalletTest {
